assignment-3/pneumonia-detection/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1
tqdm>=4.62.0
numpy>=1.21.0